+++ internal/pkg/api/repository/testresult/mongo_test.go
@@ -94,7 +94,7 @@ func TestTestExecutionsMetrics(t *testing.T) {
 	t.Run("limit should limit executions", func(t *testing.T) {
 		metrics, err := repository.GetTestSuiteMetrics(context.Background(), testName, 1, 100)
 		assert.NoError(err)
-		assert.Equal(int32(1), len(metrics.Executions))
+		assert.Equal(1, len(metrics.Executions))
 	})
 
 	t.Run("filter last n days should limit executions", func(t *testing.T) {
